[chore] #11 data class 분리
