+++ app/src/main/java/com/sopt/geonppang/domain/model/BreadType.kt
@@ -1,0 +1,8 @@
+package com.sopt.geonppang.domain.model
+
+data class BreadType(
+    val isGlutenFree: Boolean,
+    val isVegan: Boolean,
+    val isNutFree: Boolean,
+    val isSugarFree: Boolean
+)
