+++ app/src/main/java/com/sopt/geonppang/domain/model/Profile.kt
@@ -1,0 +1,7 @@
+package com.sopt.geonppang.domain.model
+
+data class Profile(
+    val memberNickname: String,
+    val mainPurpose: String,
+    val breadType: BreadType
+)
